circle

+++ rae70_Assignment1/src/rae70_assignment1/Pythagorean_Theorem.java
@@ -8,8 +8,8 @@ public class Pythagorean_Theorem {
 		String a = JOptionPane.showInputDialog("Enter side a: ");
 		String b = JOptionPane.showInputDialog("Enter side b: ");
 		//Messages that appear to user - #step 1
-			double side1 = Integer.parseInt(a);
-			double  side2 = Integer.parseInt(b);
+			double side1 = Double.parseDouble(a);
+			double  side2 = Double.parseDouble(b);
 			//^^change string to numbers #step 2
 			double pthag = Math.sqrt(Math.pow(side1,2) + Math.pow(side2,2));//Take the sqroot of a^2 + b^2 #step 3
 			double x = (pthag) * 100;
